Fix N+1 bug with characters and leader on rankings
Why:
Characters and team leader were getting loaded per team on the
leaderboard which is quite slow
*

This change addresses the need by:
Updated ranking SQL to use ActiveRecord Query methods so characters and
user can be included. Moved rank logic to build so it'll use rank if
present and fetch from DB if not. Should only have to fetch from DB on
profile team page.
*

Side effects:
Have some logic in the view that I'd rather move to a presenter or
decorator
*