Stop showing edit button on other's team profile
Summary
-------
Removed the button from showing up on other's team profile page

Why
---
The button was showing up on the team profile page giving the impression
that any team was editable. This was not the case and would just save
the characters to your team.

This change addresses the need by
---------------------------------
Check to make sure the team id matches the user's team id when on a
profile page

Side effects
------------
No way for an admin user to edit another team